# lib/api-zod/src/generated/types/currentAddress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Samaj Family Directory API
 * OpenAPI spec version: 0.1.0
 */

export interface CurrentAddress {
  current_house_no?: string | null;
  current_area?: string | null;
  current_landmark?: string | null;
  current_city?: string | null;
  current_district?: string | null;
  current_pincode?: string | null;
}