# lib/api-zod/src/generated/types/leader.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Samaj Family Directory API
 * OpenAPI spec version: 0.1.0
 */

export interface Leader {
  id: string;
  name: string;
  role: string;
  mobile?: string | null;
  address?: string | null;
  order: number;
}