# lib/api-zod/src/generated/types/userInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Samaj Family Directory API
 * OpenAPI spec version: 0.1.0
 */
import type { UserInfoRole } from "./userInfoRole";

export interface UserInfo {
  id: string;
  username: string;
  role: UserInfoRole;
}